requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
black==23.1.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.1.6
django-celery-beat==2.4.0
django-rest-framework==0.1.0
django-timezone-field==5.0
djangorestframework==3.14.0
dnspython==2.3.0
drf-spectacular==0.25.1
eventlet==0.33.3
flower==1.2.0
greenlet==2.0.2
humanize==4.6.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
kombu==5.2.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tornado==6.2
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
